Modify the logic for checking duplicate contacts
Two contacts are duplicate if the names of the two contacts are the
same.

This restricts the contacts that can be added, as it is possible for two
people to have the same name, but the system does not allow duplicate
contacts to be added.

Knowing that both the name and phone number of one contact is the
same as another contact allows us to know for a fact that the two
contacts are indeed duplicates.

Let's add an additional constraint to check if the two phone numbers are
the same.