Add restrictions when changing Role in the Edit Command
There are no restrictions or conditions when users edit the role of a
contact.

This causes student-related information and parent-related information
to be missing when the role is changed from a Staff to a Student or
Parent.

Let's add conditions for some prefixes to be present or absent when the
user edits the role of a contact.